backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1
flask-cors==4.0.1
python-dotenv==1.0.1
google-genai==1.67.0
sqlalchemy==2.0.35
